Feat/prod cpc 1038 apply products quantity section (#839)
**JIRA:** link :
https://champlainsaintlambert.atlassian.net/browse/CPC-1038

## Context:
What i completed in this ticket is a full stack with 2 new http request
both of them patch requests. When clikcing on the little plus button
which only can do he can increase or decrease the amount of each
specific product. All items below 10 productQuanity should have a css
style that mentions the item is low stock.

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
## Before and After UI (Required for UI-impacting PRs)

## Before UI


![image](https://github.com/user-attachments/assets/504a6e41-f193-49e6-b218-88b500e817bc)


## After UI 

You will see that when a product has moved from 6 to 12 quantity the CSS
will disappear. This happens after refreshing the page.


>[!Warning]
Please Note that some features such as trending and filters and maybe
more might not be working. This is due of a previous merge and we are
fixing the bugs right away



![image](https://github.com/user-attachments/assets/5836653f-22ec-4f42-8eed-3262cfc336c6)


If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links