bug/PROD-CPC-1283_RequestCount_when_create_Product_is_Null (#862)
[**JIRA:** link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1283)
## Context:
The issue started in the RecentlySeen feature. When a product was added
and clicked, a 500 error occurred due to the requestCount being null for
new products. The requestCount function would increment by one when a
newly added product was clicked, but would fail because the value would
be null. To fix this issue I started by initializing the requestCount to
0 if it's null, to make sure the increment can occur without causing an
error.

## Does this PR change the .vscode folder in petclinic-frontend?:
No changes were made to the .vscode folder.

## Changes

- Added logic to handle null values in requestCount by initializing it
to 0 if null.

## Before and After UI (Required for UI-impacting PRs)
Before: 

![image](https://github.com/user-attachments/assets/6f09e66b-87c7-41fb-b7a8-87e79546f3b1)
Error: 

![image](https://github.com/user-attachments/assets/af5eb03c-9455-4478-9fd4-f4f45dcacfb4)
Fixed:

![image](https://github.com/user-attachments/assets/ab5942f8-6f5f-43f3-a5b1-9d499c27e7fe)

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links