feat(VIST-1235): Implement Send Email Page (#868)
[**JIRA:** link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1235)
## Context:
This pull request add a new endpoint for the emailing system! This
endpoint allows sending email without using a template, which can be
useful. I also added a new button to the emailing tab of the front end!
It allows for the administrator to send email himself
## Does this PR change the .vscode folder in petclinic-frontend?:
No changes
## Changes
- New endpoint on emailing service + Test
- New endpoint on API gateway + Test
- No front end interface
## Before and After UI (Required for UI-impacting PRs)
![Screenshot 2024-10-06
000857](https://github.com/user-attachments/assets/2bfc5b55-2c28-400d-9894-3a330ee7f2f4)

![image](https://github.com/user-attachments/assets/7bbe4a07-9e71-4cef-bef3-0547b2671d7b)

![image](https://github.com/user-attachments/assets/30e3135a-0f27-4e80-93b8-61532f4ad3e4)

![image](https://github.com/user-attachments/assets/33fa701f-9416-426a-96ef-f0be350eae76)