bug/(PROD-CPC-1261): Rating not working properly (#912)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1261)
## Context:
The ticket addresses an issue where the rating functionality does not
work properly. Specifically, when a user attempts to add a rating with
any role other than an Administrator, a 403 error is thrown, but the
rating is still saved. This PR is being made to resolve the issue so
that the rating functionality works as expected for all roles, not just
administrators.
## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes

- Fixed the issue causing a 403 error when adding a rating as a
non-administrator user.
- Added the right permission handling for different user roles when
submitting ratings.

## Before and After UI (Required for UI-impacting PRs)
Before: 

![image](https://github.com/user-attachments/assets/1a677d7b-5c8c-47f4-b0e6-a4becb86e457)

![image](https://github.com/user-attachments/assets/bb67017a-9e1e-4944-a5ee-b36d642ec933)
After:

![image](https://github.com/user-attachments/assets/7cba6223-5742-49e5-b0c9-49ade4afd2be)

## Dev notes (Optional)
Adjusted backend role handling when using getProductByProductId.
## Linked pull requests (Optional)
Pull request links

Co-authored-by: Viktor Kuts <Vik.Kuts@hotmail.com>