feat(VETS-CPC-1388): rework vet page ui (#927)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1388
## Context:
On the main page of the vet, we only had a table with info such as first
and last and speciality. We didn't have the picture of the vets and the
page wasn't presentable. Modifications were made to have a cleaner and
presentable vet page. Instead of a table we now have cards that display
the first and last name, the vet specialities and their photo.

## Does this PR change the .vscode folder in petclinic-frontend?:
This PR does not affect the .vscode file

## Changes

- Removed the vet table.
- Added cards instead of the table.
- Added the vet photos in the cards on top of the first and last names
and specialities.

## Before and After UI (Required for UI-impacting PRs)
Before:

![image](https://github.com/user-attachments/assets/0c62fcd0-d572-4cee-9e40-2e17cf134cd5)

After:

![image](https://github.com/user-attachments/assets/d016a3b6-35b5-45b7-ab70-858acfd9295c)

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links