feat(VIST-CPC-1319): View A Specific Emergency Visit (#932)
https://champlainsaintlambert.atlassian.net/browse/CPC-1319
## Context:
What is the ticket about and why are we doing this change.
The ticket is to add the functionnality to see a specific emergency
Visit. I do this change so a user can see all the details of his visits
with more simplicity.
## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not. No
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
- Needed to update the getEmrgencyVisit in the backend
- Implemented the page ui in the front end
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

![image](https://github.com/user-attachments/assets/48dea11b-9a8a-4137-b46f-f2f0247b81b0)

![image](https://github.com/user-attachments/assets/0f65029c-174f-4108-b712-fe4b8abdfa63)



## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links

Co-authored-by: Zakaria_ <135060272+tramway23@users.noreply.github.com>