feat(VETS-CPC-1375): update vet education information (#941)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1375
## Context:
This PR is about updating the vets education information if ever they
need to. A new button is now accessible to all vets that, when pressed,
a form with the existing vet education information will open, and vets
will be able to modify it as they like.

## Does this PR change the .vscode folder in petclinic-frontend?:
This PR does not effect the .vscode file.

## Changes

- Added the update education method in the back-end for the v2 API.

- Created tests for the method.

- Added the whole front-end logic to be able to make modifications on
vet education info.

## Before and After UI (Required for UI-impacting PRs)
Before:
There was no button below each education information. 

![image](https://github.com/user-attachments/assets/62c938c2-06bd-4270-a3ea-552a2f570b1e)


After: 

![image](https://github.com/user-attachments/assets/23ab3caa-e4ef-4a10-ba0d-3dbc951917e7)

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links