bug/(PROD-CPC-1414): Fix Null Average Rating Causing 500 Error on New Product Added (#959)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1414)
## Context:

This ticket fixes a bug where adding a new product without ratings
caused a 500 error due to a NullPointerException in the
getAverageRating() method. The issue was that the averageRating was set
to null, which caused problems with sorting and filtering. The fix now
sets averageRating to 0.0 by default
## Does this PR change the .vscode folder in petclinic-frontend?:
No, this PR does not change anything in the .vscode folder.

## Changes

- Updated getAverageRating() method in ProductServiceImpl.java to set
averageRating to 0.0 when there are no ratings.
- Fixed tests that were causing issues.

## Before and After UI (Required for UI-impacting PRs)
Before:

![image](https://github.com/user-attachments/assets/581c3f27-6506-4b19-a791-8c8371d244e4)

After:

![image](https://github.com/user-attachments/assets/ebc80464-ce01-42e6-a1ad-e2e7db17e5ed)

---------

Co-authored-by: Jessy Gjerek <118135201+J91-cloud@users.noreply.github.com>