bug(INVT-CPC-1417): make uploaded images not required (#987)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1417
## Context:
Everytime we want to update our inventory we must always select an image
because our uploaded image input box requires a value and it doesn't
already have a value. This is because logic isn't implemented to fetch
the uploaded image from the database and store that value fetched into
the input box. But this isn't needed thanks to this simple fix.
## Does this PR change the .vscode folder in petclinic-frontend?:
**No, it doesn't**
## Changes
• Removed required from both the add inventory form and edit inventory
form
## Before and After UI (Required for UI-impacting PRs)
## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links