feat(INVT-CPC-1153): Searching products status (#1001)
**JIRA:** link to jira ticket
## Context:
Before, the front-end was capable of filtering the status of the
products of an inventory, but not the back-end. Now, the back-end method
meant to search products inside of an inventory is supporting the
filtering by name, description and status of products. This change was
necessary for later development where the search product feature will be
exploited outside of the current inventory products page.
## Does this PR change the .vscode folder in petclinic-frontend?:
No.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
Back-end method meant to search products inside of an inventory is
supporting the filtering by name, description and status of products.
Increase testing coverage in the inventory-service and the api-gateway
for the searchProducts method.