bug/(VETS-CPC-1418): fix code for get photos (#1010)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1418)

## Context:
This PR is about changing the way the API call was coded in the
VetDetails page. We have a lot of API call methods in the vetDetails
page and it is bad coding practice. Each API call should be in its own
separate file and should be using Axios. So I used the API call that I
created in a previous PR and removed the one in VetDetails page.

## Does this PR change the .vscode folder in petclinic-frontend?:
This PR does not effect the .vscode file

## Changes

- Removed the API call from the VetDetails page.
- Created a method that would call the fetchVetPhoto API call from the
fetchPhoto.ts file

## Before and After UI (Required for UI-impacting PRs)
Before:

![image](https://github.com/user-attachments/assets/511ac4b0-5f57-47a3-b2c6-4f40313c1fc6)

After:

![image](https://github.com/user-attachments/assets/baae080e-3d9b-417d-911a-3716058caebe)
 

## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links