bug/(PROD-CPC-1262): Misuse of the word product (#1009)
**JIRA:** [link to jira
ticket](https://champlainsaintlambert.atlassian.net/browse/CPC-1262)
## Context:
This ticket fixes the issue where "Products" is used everywhere in the
shop. The change will make "Product" refer to items and "Shop" refer to
the whole catalog. This improves clarity for users.
## Does this PR change the .vscode folder in petclinic-frontend?:
No, this PR does not make any changes to the .vscode folder.

## Changes

- Changed the naming used with products to items 
- Changed the naming used for better readability and clarity for users
in several places in the Shop page

## Before and After UI (Required for UI-impacting PRs)

Before: 

![image](https://github.com/user-attachments/assets/b15531e5-8a02-41fa-b239-224ae8eb66a0)

After:

![image](https://github.com/user-attachments/assets/b7c3c9ab-f089-4223-b924-635e1a122f49)