config/e2e playwright initial setup (#1023)
**JIRA:** 
none
## Context:
Added basic configuration for playwright to be used as an e2e testing
tool. It is not yet connected to the CI pipeline.
## Does this PR change the .vscode folder in petclinic-frontend?:
Yes. I added the playwright extension to extensions.json file.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
- updated tsconfig.json to work with playwright tests
- added playwright test for vscode
- updated default playwright.config.ts 
- added various test files containing example test cases for petclinic

## Before and After UI (Required for UI-impacting PRs)
n/a 
## Dev notes (Optional)
This is a very basic installation. In the future, it needs to be updated
with the following:
- automatic login of admin and other roles
- preset locators for all the different pages
- anything to make the tests simpler and more fool-proof to create

## Linked pull requests (Optional)
none