e2etest/PROD-Zachary (#1056)
**JIRA:** [Link to JIRA ticket]

## Context:
This ticket focuses on adding a test to ensure that the review filter
works correctly when users set a minimum rating equal to or higher than
the maximum rating. The test verifies that an error message is displayed
if this condition is met.

## Does this PR change the .vscode folder in petclinic-frontend?:
No changes were made to the `.vscode` folder. Reviewers, please confirm
that no unintentional changes were made to this folder.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
- Added validation to the review filter for minimum and maximum star
ratings.