bug/(BILL-CPC-1440): Fix_Improper_Use_Of_API_Calls_In_FE (#1068)
…of hardcoded urls.

**NOTE
- The post request on the admin page to create a new bill creates a new
bill with the first name being null, I will create a ticket on jira to
fix this issue
- The get by billId is not working, I will create a ticket on jira to
fix this issue
- I changed the getBillsByCustomerId method in the
CustomerBillController to return APPLICATION_JSON_VALUE, since I
couldn't get the front-end to handle SSE with axios on this one. I will
create a ticket on jira to fix this issue

**JIRA:** link to jira ticket
## Context:
The API's being hardcoded causes the hosted front-end to not work.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
- The front-end API calls now use the axios format provided by Dylan.
## Does this use the v2 API?:
Yes because some of the requests in the front-end were using v2
## Does this add a new communication between services?:
NO
## Before and After UI (Required for UI-impacting PRs)
No changes
## Dev notes (Optional)
The getBillsByCustomerId in CustomerBillController was changed from
text-event stream to application json because I tried to make it handle
SSE but failed and would rather get my changes up on main as soon as
possible to get on to my next task