feat (PROD-CPC-1472): change recently seen to be horizontally scrollable (#1076)
**JIRA:** link to jira ticket
## Context:
What is the ticket about and why are we doing this change.

This ticket was to make the recently viewed section horizontally
scrollable to create a better UI for the user.
## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

The changes I made did not change any other modules. I added css and
changed the recently seen to be flex instead of a grid. The changes were
made in the ProductList.css, ProductList.tsx, and
RecentlyViewedProducts.tsx files.
## Does this use the v2 API?:
If the PR uses the v2 API, explain why
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
## Before and After UI (Required for UI-impacting PRs)
<img width="1469" height="818" alt="Screenshot 2025-09-15 at 3 45 14 PM"
src="https://github.com/user-attachments/assets/ff44ea5c-bf74-4372-9dd0-db8542666e1a"
/>



<img width="1002" height="533" alt="Scroll"
src="https://github.com/user-attachments/assets/390f469a-2f59-4900-90dc-10d5fcc05e83"
/>


If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links