bug(VETS-CPC-1279) : uplaod photo vetId dropdown menu (#1067)
**JIRA:** https://champlainsaintlambert.atlassian.net/browse/CPC-1279
## Context:
This ticket is about the upload vet photo modal where in order to choose
what vet to upload the photo for you had to manually type in the VetId
into a text field. This is not user friendly and was therefore replaced
with a dropdown menu that listed the names of the vet, and would still
pass the vetId where it was needed, so that it would be easier to upload
a photo for a vet.
## Does this PR change the .vscode folder in petclinic-frontend?:
No, it does not change the .vscode folder.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
The only files affected by this change, given this is only a UI fix, are
the Vet.tsx and the UploadVetPhoto.tsx files. In Vet.tsx I passed the
allVets as a prop into UploadVetPhoto so that it could use it when
creating the dropdown menu and to get the ids of each vet.
## Does this use the v2 API?:
No.
## Does this add a new communication between services?:
No.
## Before and After UI (Required for UI-impacting PRs)
Before:
<img width="2556" height="1357" alt="before"
src="https://github.com/user-attachments/assets/e3fa2db5-b74c-4d7f-9739-e9b0e383a004"
/>

After:
<img width="2553" height="1283" alt="after"
src="https://github.com/user-attachments/assets/5eac8762-67c4-4ac3-90fe-1703323ec68b"
/>