feat(CUST-CPC-665): Full integration of DeletePetType  (#1115)
**JIRA:** link to jira ticket

https://champlainsaintlambert.atlassian.net/jira/software/c/projects/CPC/boards/14?label=CUST&selectedIssue=CPC-665
## Context:
This ticket addresses the issue that a logged in admin is not able to
delete a pet type via the front-end. After implementing the feature we
realized that the back-end was not really working and was mostly
deprecated. Which led me to implement a new feature (delete pet type),
for both front end and back end.
## Does this PR change the .vscode folder in petclinic-frontend?:
No it does not modify the .vscode folder

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

- Fully implementing the deletePetType on the Angular front-end
- Changed the API-Gateway method for deletePetType (V2 Creation)
- Changed various other instances of the deletePetType since the other
version was unworkable/deprecated

## Does this use the v2 API?:
I created a V2 since the V1 was not returning the proper information, it
was also not linked to the customer-service-reactive. It was working as
a standalone and was deprecated. I also changed the backend code to be
able to correctly implement the V2
## Does this add a new communication between services?:
No it does not
## Before and After UI (Required for UI-impacting PRs)
<img width="2560" height="923" alt="Capture"
src="https://github.com/user-attachments/assets/9f012357-6a72-433e-84b1-d2d1ec8cf914"
/>
<img width="2560" height="920" alt="Capture1"
src="https://github.com/user-attachments/assets/c4dd7c3e-01f8-418c-b810-8074d2fb12ce"
/>

## Dev notes (Optional)
No
## Linked pull requests (Optional)
No

---------

Co-authored-by: Mir Faiyazur Rahman <137716472+Mirf-Rahman@users.noreply.github.com>