feat(AUTH-CPC-1522): Remove new emailing service implementations and improve the design of the old mailing service (#1138)
**JIRA:**
[link](https://champlainsaintlambert.atlassian.net/browse/CPC-1522?atlOrigin=eyJpIjoiMzBkOTllZDFhNzdiNDhmNThmOTg0YmU3MTFhM2E2ZGQiLCJwIjoiaiJ9)
## Context:
Change all the uses of the new emailing service with the go mailing
service, as well as improve the design of the old mailing service. This
way we keep things simple easy to understand.
## Does this PR change the .vscode folder in petclinic-frontend?:
No
<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
-Changed the old mailing service's sendEmail method in the UserService
and UserServiceImpl
-Added new dto for old mailing service to base everything off of
-Removed new emailing service's methods in UserService and
UserServiceImpl
-Refactored the tests in the auth service for the old mailer service
that were affected by dto change
-Removed the new emailing service's tests
-Deleted EmailingServiceClient(new emailing service)
-Changed the old mailing service struct to match the new dto in the auth
service
-Refactored the files in old mailing service folder in auth service to
work with new dto
-Refactored the code in the controller, middleware, service, and mailer
in the old mailing service to work with new struct
-Refactored all the tests for the go service to match with the new
struct
## Does this use the v2 API?:

## Does this add a new communication between services?:
no
## Before and After UI (Required for UI-impacting PRs)
Did not touch front end
## Dev notes (Optional)
I needed to change go version in from 1.20 to 1.25(1.24 should work) for
some packages that are imported in the service. I will change it back in
the PR. So if the go packages are being found when you try to build the
project, change the version from 1.20 to 1.24 or newer.
## Linked pull requests (Optional)