bug/BILL-CPC-1568-Fixing-Delete-Method (#1152)
Link to Jira ticket:
https://champlainsaintlambert.atlassian.net/browse/CPC-1568

### Context:

-----------------------------------------------------------------------------------------------------------------------------------------------
What/Why: This ticket is about fixing the delete method so it can allow
the deletion of bills and adding one negative delete test. This needed
to be fixed because not being able to delete bills causes an overflow of
data, it also keeps the page nicer and organized. The negative delete
test needed to be added just to test when deleting a bill that does not
exists it would return not found.

### Does this change affect the .vscode folder in the front-end:

-----------------------------------------------------------------------------------------------------------------------------------------------
No

### Changes:

-----------------------------------------------------------------------------------------------------------------------------------------------
Changed the Uri in the pet-clinic front-end application in the
DeleteBill.tsx to match the Uri from the bill controller class. Added
one negative delete test for returning a bill not found.

### Does This use the v2 API:

-----------------------------------------------------------------------------------------------------------------------------------------------
No

### Does this add a new communication between services:

-----------------------------------------------------------------------------------------------------------------------------------------------
No

### Before/After Ui changes:

-----------------------------------------------------------------------------------------------------------------------------------------------
N/A