Bug(AUTH-CPC-1578): reduce memory usage on spring boot docker containers (#1155)
**JIRA:** link to jira ticket
[link](https://champlainsaintlambert.atlassian.net/browse/CPC-1578?atlOrigin=eyJpIjoiMGQwMWZhM2ZkZTRkNGNhMjg4NzI5MmVjMjY0YmQ3MjUiLCJwIjoiaiJ9)
## Context:

The Auth Service, as well as many others use too much ram
The goal is to try to minimize the amount of memory used on the auth
service, so we know what to do for all subsequent services.

## Does this PR change the .vscode folder in petclinic-frontend?:
no

## Changes
- changed the auth service dockerfile to use a distroless jdk, reducing
storage bulk and ram usage.
- started a Prometheus+Grafana stack for monitoring things like JVM
memory usage and garbage collection to help

## Does this use the v2 API?:
no changes to api calls

## Does this add a new communication between services?:
At the moment, monitoring is optional and not included in the main
docker-compose

## Before and After UI (Required for UI-impacting PRs)
No UI change

## Dev notes (Optional)
It's only implemented for the auth service as we need to expose the
metrics to scrape

## Linked pull requests (Optional)
not applicable

---------

Co-authored-by: Jeremy Roos <149620614+JeremyNRoos@users.noreply.github.com>