feat(PROD-CPC-1551): Changed the front end of the products page (#1143)
**JIRA:** link to jira ticket
## Context:
What is the ticket about and why are we doing this change.
change the shop page to look similar to this example. As for the
trending and bundle leave it as is for now, unless you know a better way
to style and integrate it.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.
- I changed the design of the cards
- changed the layout
## Does this use the v2 API?:
If the PR uses the v2 API, explain why
## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.
## Before and After UI 

[FrontEndBefore.pdf](https://github.com/user-attachments/files/22571104/FrontEndBefore.pdf)
(Required for UI-impacting PRs)

<img width="959" height="562" alt="Screenshot 2025-09-28 171857"
src="https://github.com/user-attachments/assets/d3584555-8e0a-45f5-be27-2b7226210eef"
/>
<img width="1006" height="563" alt="Screenshot 2025-09-28 171943"
src="https://github.com/user-attachments/assets/48c8e5ff-3000-4812-a09a-73d89d3454ee"
/>
<img width="1018" height="518" alt="Screenshot 2025-09-28 172012"
src="https://github.com/user-attachments/assets/dbb90914-34aa-4570-9804-06c40fa62231"
/>



These don't really show how the page looked before and how it looks now.


If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.
## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links