feat/(PROD-CPC-1596): Add controller v1 and its test for products (#1177)
**JIRA:** 

https://champlainsaintlambert.atlassian.net/browse/CPC-1596?atlOrigin=eyJpIjoiNTg4YzRhZDgzNGYyNGIwZWE5NmFlMGFlNGNhZmJlMTgiLCJwIjoiaiJ9

## Context:
So far we don't have any v1 so we wanted to move everything form v2 to
v1 and having tests for the v1 that we created in the api-gateway

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes

- Added ProductControllerV1
- Added tests (ProductControllerV1IntegrationTest and
ProductControllerV1UnitTest)
- Re-organize, put v2 tests in a Products package

## Does this use the v2 API?:
Everything is currently still using v2 for products

## Does this add a new communication between services?:
If the PR adds a new communication between services, make sure to update
the C4 L2.

## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.

## Dev notes (Optional)
<img width="1399" height="724" alt="image"
src="https://github.com/user-attachments/assets/7aee04bc-a20b-4105-a0a5-4fa91584248b"
/>
<img width="488" height="511" alt="image"
src="https://github.com/user-attachments/assets/52d90da9-c430-4862-8467-a55afaf06a2f"
/>



## Linked pull requests (Optional)
Pull request links

Co-authored-by: Seddik Ayoub <102536731+Ayoub-Seddik@users.noreply.github.com>