feat(VIST-CPC-1219): fix edit emergency (#1183)
**JIRA:** [feat(VIST-CPC-1219): fix edit emergency:
put](https://champlainsaintlambert.atlassian.net/browse/CPC-1219?atlOrigin=eyJpIjoiYjU0Njc4Mjc0OWIzNDljMjkyZmY1OTQ5MThiMTA3OGMiLCJwIjoiaiJ9)
## Context:
What is the ticket about and why are we doing this change.

The "Edit" button in the Emergency table only opens the view page for
that Id. It doesn’t show a form for editing.

## Does this PR change the .vscode folder in petclinic-frontend?:
If the PR changes the .vscode folder, explain why in detail because it
should not.
Be sure to include cgerard321 as a reviewer.


No. It does not. 

<span style="color:red">**Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review
comments.**</span>

## Changes
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.


- Added "@JsonFormat(pattern = "yyyy-MM-dd HH:mm")" to the EmergencyDTOs
This was added to ensure the visitDate format stays consistent. 

Made changes in the following files:

.java

- VisitsServiceClient.java
- EmergencyRequestDTO.java
- EmergencyResponseDTO.java
- VisitController.java
- EmergencyService.java
- EmergencyServiceImpl.java
- EmergencyRequestDTO.java
- EmergencyResponseDTO.java
- VisitController.java

.css

- AddEmergencyForm.css

.tsx

- AddEmergencyForm.tsx
- EditEmergency.tsx
- VisitListTable.tsx
- router.tsx

.ts

- updateEmergency.ts
- EmergencyRequestDTO.ts
- EmergencyResponseDTO.ts
- path.routes.ts


## Does this use the v2 API?:
If the PR uses the v2 API, explain why

Yes. It uses the new version because the updateEmergency does not exist
in the BFF.

## Does this add a new communication between services?:


No. It does not. 

If the PR adds a new communication between services, make sure to update
the C4 L2.
## Before and After UI (Required for UI-impacting PRs)
If this is a change to the UI, include before and after screenshots to
show the differences.
If this is a new UI feature, include screenshots to show reviewers what
it looks like.


BEFORE: 
<img width="1906" height="578" alt="image"
src="https://github.com/user-attachments/assets/2e8fcf8e-61f1-41d8-b2ae-5295351ec557"
/>


AFTER:
<img width="606" height="779" alt="image"
src="https://github.com/user-attachments/assets/e6daf7f3-7423-436e-b5f2-12ccb02d71ef"
/>



## Dev notes (Optional)
Specific technical changes that should be noted
## Linked pull requests (Optional)
Pull request links

---------

Co-authored-by: Anvar <73323166+anvsult@users.noreply.github.com>